requirements.txt
streamlit==1.40.1
apache-airflow==2.10.5
great-expectations==1.1.0
numpy==1.24.4
pandas==2.0.3
fastapi==0.115.8
uvicorn==0.33.0
python-dotenv==1.0.1
pytest==8.3.4
ruff==0.9.4
pytest-asyncio==0.24.0
xgboost==2.1.3
python-multipart==0.0.20
scikit-learn==1.3.2
sqlmodel==0.0.11
psycopg2-binary==2.9.10
pymsteams==0.2.4
alembic==1.14.1
apache-airflow-providers-postgres==5.13.0
requests==2.32.3
pyroscope-io==0.8.8
chardet==5.2.0
joblib==1.4.2